feat: use asyncio to make adding photos faster
Adding photos used to take a considerable amount of time. For
each PHAsset, the image was copied, saved, and the header view
reset. The most obvious time-saver was to only set the header
if it was previously blank (i.e. default cactus). The saving
of the photos is now completed in the background.

Minor feature still being worked on is loading the images for
the library table view in the background. Cropping and resizing
takes considerable amount of time, slowing the navigation from
detail to Library table view. This has been fixed, but now the
images do not appear upon loading the app until the cell is
selected or recycled (re-enters the view).