feat: load images with async upon first load
TLDR: I fixed the bug that caused no image to be shown in the
Library's cells upon the first loading.

Soln: Since the cell was returned with `imageView.image = nil`
in `cellForRowAt`, the view did not allocate viewing space in the
cell for an image. Therefore, when a UIImage was sent, it was not
visible. However, when the cell is recycled, an image already
existed, so the space was made for the imageView and the old
image was replaced with the new one when async caught up.

Fix: Include a small "blankImage" asset that is cropped, resized,
and loaded if `imageView.image == nil`. This reserves the space
and is quick for the app to load.

Long-term fix: save the cropped, resized image for each Plant
object and use that if it exists.