feat: paging image viewer, fix: save Plants after adding images
paging image viewer:
- swipe to advance images
- tap to enter black background, hide nav bar and status bar

save Plants after adding images
Problem: the Plants were not saved after saving images picked
in the detail view controller. If the user saved images and
then closed the app before returning to the table view, the
Plant object was not saved, and the new image URLs were lost.

Soln: create a protocol wtih `savePlants()` method, set the Library
as the delegate for the detail view controller (weak to avoid
strong reference cycles)

Minor issue: have to call it on every image because the images are
save asynchronously. Issue opened on GitHub.