fix: properly save and load saved images
Problem: After every rebuild of app, image paths no longer worked.

Reason: On each new build, the Documents dir moves under a new
UUID.

Solution: Save the UUID of the image and append it to the
Documents directory for each build. (Pretty simple, but it took
awhile for me to realize this -_-.)