refactor: tried to use SwiftPhotoGallery but reverted back
For all of the features it purported to have, they were
more of a nusiance than anything else.

I don't think it will be trivial to get pinch-to-zoom in the
paging photo viewer, but I think I can figure it out. The zoom and
fade transtions will be more difficult, but still do-able.