feat: can zoom into images in paging image view
Reformatted the framework to use a collection view that scrolls
horizontally instead of a scroll view. I also made a specific
class of collection view cell. This has the image view inside of a
scroll view which enables control over zooming into an image.

Still small UI components to account for, but the framework is
mostly there.