refactor: PlantLibraryTableView refactored
Biggest change is to use a custom table view cell:
PlantLibraryTableViewCell. It has a stored property for a Plant
object and then uses that to update the labels and icon.

Main function for new cell is `cell.setupCell()` which calls
`self.setupConstraints()` and then `self.setupCellView()`.

It is conected to `PlantLibraryTableViewCell.xib`.