feat: tap on image in paging view to make background black/white
Also hides and shows the navigation bar. Annoyingly, I had to
implement the tag recognizer in the table view cell, then used
a delegate to control the navigation bar in the view controller.

While this seems to make the cell, a view, do too much, because
it does not affect the model, I see no reason to need to make
this more complicated by introducing a controller for the cell.
This is user interaction with the view ONLY, not with the model.