feat: import images in image collection; refactor of delegates
I refactored the delegate system for the heirarchy of views.
Instead of sending a Plant delegate and only keeping the plant
object in the detail view controller, the actual plant object is
passed to the various view controllers.

Can now import images from library in the image collection view.
The above refactoring was necessary for this.