fix: paging collection view black/white on tap fixed
Originally, the tap gesture would turn the cell background black
and tell the collection view controller to hide the nav bar using
a delegate system. The problem was that only that cell, and not
those adjacent, were black. Therefore, the delegate was put in
charge of changing the collection view's background, too. This
resulted in a simpler implementation, anyways.