feat: finished plant level editing system
There was one more bug to address that caused weird issues when
"moving" the editing cell. I think the issue was that the cell
being removed and inserted was the same object. Changing the edit
manager's system to address this hypothesis fixed the problem. Now
the edit manager's editing cell is a new object whenever the plant
level being edited it changed (a didSet property for the plantLevel
value).